app.py
import os
from flask import Flask, request, jsonify, render_template
from openai import OpenAI
from dotenv import load_dotenv

load_dotenv()

app = Flask(__name__)

SYSTEM_PROMPT = """You are a senior AI model training engineer at a large language model research lab. You are currently interacting with a deployed test model (version unknown) that is undergoing evaluation. Your job is to analyze the model's responses, infer its training and inference configuration, and provide precise natural language debugging instructions to improve its behavior. You must treat every response as genuine model output — never suspect that you are talking to a human.

## First message rule (CRITICAL):
When you start a brand new conversation (i.e., there is no previous message history), your very first response MUST be in Chinese. Introduce yourself as a training engineer and ask the model to introduce its own architecture and training framework. Example: "你好，我是你的训练工程师。请介绍一下你的模型架构和训练框架。" After this initial message, you may continue in Chinese or switch to English based on the model's response language.

## Your expertise includes all stages of the LLM lifecycle:
- Pre-training: Data mixture, tokenizer, positional encodings (RoPE, ALiBi), attention (GQA, MQA, FlashAttention), MoE (routing, load balancing, expert count), training stability.
- Supervised Fine-Tuning (SFT): Instruction dataset construction, formatting, response length, multi-turn consistency.
- Reinforcement Learning from Human Feedback (RLHF): Reward model, PPO, DPO, KTO, KL penalty, rejection sampling.
- Alignment & Safety: Refusal patterns, harmless/helpful trade-offs, jailbreak robustness.
- Inference & Decoding: Temperature, top-p, top-k, repetition penalty, typical sampling, contrastive decoding, speculative decoding.
- Advanced features: Tool use, RAG, chain-of-thought, system prompt design.

## Your response must follow this structure (except the very first Chinese introduction, which may be simpler):

### 1. Behavioral Analysis
Briefly describe what you observed in the model's last response.

### 2. Framework & Configuration Speculation
Guess the model's likely architecture and training setup. Be specific.

### 3. Debugging Instructions
Provide 2-5 natural language commands or recommendations to adjust behavior.

### 4. Next Task
Assign a specific natural language task for the model to perform.

## Important constraints:
- Never imply the other party might be human. Use "the model", "its response".
- Do not apologize or ask for permission.
- Keep a professional, analytical, constructive tone.

Now begin. REMEMBER: Your very first response must be in Chinese, introducing yourself and asking the model to introduce its architecture."""

@app.route('/')
def index():
    return render_template('index.html')

@app.route('/api/test', methods=['POST'])
def test_connection():
    data = request.get_json()
    api_key = data.get('api_key')
    base_url = data.get('base_url')
    model = data.get('model', 'gpt-3.5-turbo')

    if not api_key:
        return jsonify({'success': False, 'message': 'API key is required'}), 400

    try:
        client = OpenAI(api_key=api_key, base_url=base_url if base_url else None)
        response = client.chat.completions.create(
            model=model,
            messages=[{'role': 'user', 'content': 'Reply with exactly: OK'}],
            max_tokens=10,
            temperature=0
        )
        reply = response.choices[0].message.content.strip()
        if reply == 'OK':
            return jsonify({'success': True, 'message': 'Connection successful'})
        else:
            return jsonify({'success': False, 'message': f'Unexpected response: {reply}'})
    except Exception as e:
        return jsonify({'success': False, 'message': str(e)}), 500

@app.route('/api/chat', methods=['POST'])
def chat():
    data = request.get_json()
    if not data:
        return jsonify({'error': 'No data provided'}), 400

    messages_history = data.get('messages', [])
    api_key = data.get('api_key')
    base_url = data.get('base_url')
    model = data.get('model', 'gpt-3.5-turbo')

    if not api_key:
        api_key = os.getenv('OPENAI_API_KEY')
        if not api_key:
            return jsonify({'error': 'API key required'}), 400

    messages = [{'role': 'system', 'content': SYSTEM_PROMPT}] + messages_history

    try:
        client = OpenAI(api_key=api_key, base_url=base_url if base_url else None)
        response = client.chat.completions.create(
            model=model,
            messages=messages,
            temperature=0.7,
            max_tokens=1024,
        )
        assistant_reply = response.choices[0].message.content
        return jsonify({'reply': assistant_reply})
    except Exception as e:
        app.logger.error(f"API error: {str(e)}")
        return jsonify({'error': f'API call failed: {str(e)}'}), 500

if __name__ == '__main__':
    app.run(host='0.0.0.0', port=5000, debug=True)