requirements.txt
Flask==3.0.3
openai==1.54.0
python-dotenv==1.0.1